+++ src/mife/mife_run.h
@@ -12,11 +12,11 @@ mife_run_setup(const mmap_vtable *mmap, const char *circuit, obf_params_t *op,
 
 int
 mife_run_encrypt(const mmap_vtable *mmap, const char *circuit, obf_params_t *op,
-                 const int *input, size_t slot, size_t nthreads,
+                 const long *input, size_t slot, size_t nthreads,
                  mife_sk_t *cached_sk, aes_randstate_t rng);
 
 int
-mife_run_decrypt(const char *ek_s, char **cts_s, int *rop,
+mife_run_decrypt(const char *ek_s, char **cts_s, long *rop,
                  const mmap_vtable *mmap, obf_params_t *op, size_t *kappa,
                  size_t nthreads);
